refactor: move WebviewUpdater to managers
